requirements.txt
flask
flask-cors
google-genai>=1.73.1
python-dotenv
